Update launching_instance.md
Why Step 3b requires the instance to be deployed to a public subnet? Why is internet connectivity required if the AWS Managed AD is deployed in private subnets?
"If this is not the case, you won't be able to connect to the instance remotely\." This sentence is misleading, because it is possible to RDP in the instance even if it is deployed in private subnet by using bastion hosts, VPN, Direct Connect, etc.
It would be better to rephrase Step 3b and clarify what is required and why.